% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{app_launcher}
\alias{app_launcher}
\title{RPLANES explorer app launcher}
\usage{
app_launcher(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \link[shiny:runApp]{shiny::runApp}}
}
\value{
This function starts the Rplanes Shiny app.
}
\description{
The Rplanes explorer app allows a user to interactively use their own data sets, or an internal example data set, to explore the various functions developed for plausibility analysis.
}
\examples{
\dontrun{
# Launch the explorer app
app_launcher(host = "0.0.0.0",
             launch.browser = TRUE,
             port = 80)
}
}
