% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date_spans}
\alias{check_date_spans}
\title{Check date spans}
\usage{
check_date_spans(seed_date, this_date, resolution)
}
\arguments{
\item{seed_date}{fixme}

\item{this_date}{fixme}

\item{resolution}{fixme}
}
\value{
Invisible: a list of input values
}
\description{
Description FIXME
}
\details{
Details here FIXME
}
\examples{
seed_date <- as.Date("2023-03-08")
this_date <- as.Date("2023-03-15")
check_date_spans(seed_date = seed_date, this_date = this_date, resolution="weeks")
x <- try(check_date_spans(seed_date = seed_date,
                          this_date = this_date,
                          resolution="days"), silent=TRUE)
x
x <- try(check_date_spans(seed_date = seed_date,
                          this_date = this_date,
                          resolution="months"), silent=TRUE)
x
}
