% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutter}
\alias{cutter}
\title{Cut into categorical differences}
\usage{
cutter(x, threshold = 1)
}
\arguments{
\item{x}{Vector of length 1 with scaled difference to be categorized}

\item{threshold}{Limit used to define the categorical differences; default is \code{1}}
}
\value{
Character vector of length 1 with the categorical description of difference
}
\description{
This unexported helper function takes an input number for an observed difference and cuts it into a categorical description (e.g., "increase", "decrease", or "stable") of the change.
}
