% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epiweek_start}
\alias{epiweek_start}
\title{Epiweek start}
\usage{
epiweek_start(date)
}
\arguments{
\item{date}{Date to be queried}
}
\value{
Date of the first day of the epiweek for the input date.
}
\description{
This unexported helper identifies the date of the first day for the epiweek of the given date. The function is used internally inside of \link{valid_dates}.
}
