% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_observed}
\alias{is_observed}
\title{Check if object is observed}
\usage{
is_observed(x)
}
\arguments{
\item{x}{Input object to be checked}
}
\value{
Logical as to whether or not the input object inherits the "signal" and "observed" classes.
}
\description{
Check if object is observed
}
