% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%m-\%}
\alias{\%m-\%}
\title{Month subtraction}
\arguments{
\item{e1}{A period or a date-time object of class \link{POSIXlt}, \link{POSIXct}
or \link{Date}.}

\item{e2}{A period or a date-time object of class \link{POSIXlt}, \link{POSIXct}
or \link{Date}. Note that one of e1 and e2 must be a period and the other a
date-time object.}
}
\value{
A date-time object of class POSIXlt, POSIXct or Date
}
\description{
See \code{lubridate::\link[lubridate:add_with_rollback]{\%m-\%}} for details.
}
\keyword{internal}
