% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planes.R
\name{plane_score}
\alias{plane_score}
\title{Score PLANES components}
\usage{
plane_score(input, seed, components = "all")
}
\arguments{
\item{input}{Input signal data to be scored; object must be one of \link[=to_signal]{forecast} or \link[=to_signal]{observed}}

\item{seed}{Prepared \link[=plane_seed]{seed}}

\item{components}{Character vector specifying components; Any combination of "cover", "diff", "taper" and "repeats". Default is \code{'all'} and will use all available components for the given signal}
}
\value{
A \code{list} with scoring results for all locations.
}
\description{
This function wraps PLANES scoring for specified components across all locations in single step.
}
