% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane_shape_temp.R
\name{plane_shape}
\alias{plane_shape}
\title{Shape Component}
\usage{
plane_shape(location, input, seed)
}
\arguments{
\item{location}{Character vector with location code; the location must appear in input and seed}

\item{input}{Input signal data to be scored; object must be one of \link[=to_signal]{forecast} or \link[=to_signal]{observed}}

\item{seed}{Prepared \link[=plane_seed]{seed}}
}
\value{
A \code{list} with the following values:
\itemize{
\item \strong{indicator}: Logical as to whether or not the shape of the evaluated signal is novel
\item \strong{output}: An n x 5 tibble. The columns are:
\itemize{
\item \strong{date}: The dates corresponding to all observed and forecast data (formatted as date)
\item \strong{point}: The incidence of all observed and forecast data (e.g., hospitalization rates)
\item \strong{type}: Indicates whether the data row is observed or forecast data
\item \strong{indicator}: Logical as to whether or not the shape of the evaluated signal is novel
\item \strong{location}: A character vector with the location code
}
}
}
\description{
This function evaluates the shape of the trajectory of the forecast signal and compares that shape to existing shapes in observed data. If the shape is flagged as a novel shape, a flag is raised, and the signal is considered implausible. See details for further information.
}
\details{
This function uses a Dynamic Time Warping (DTW) algorithm to identify shapes within the observed seed data.

sliding windows
dtw on observed
dtw comparing observed distances to signal distances
if distance > threshold, than it is novel and it is flagged as implausible
}
\examples{
adfsadfa

}
\references{
asdfasfa
}
