% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_resolution}
\alias{resolve_resolution}
\title{Resolve resolution}
\usage{
resolve_resolution(resolution)
}
\arguments{
\item{resolution}{Character vector specifying the temporal resolution (e.g., "days", "weeks", "months")}
}
\value{
If the resolution matches to "days", "weeks", or "months" then the match will be returned. If not, the function will throw an error.
}
\description{
This helper function uses argument matching to resolve the resolution from input. The function also handles casing. This will allow, for example, an input resolution of "daily" or "day" to be resolved to "days".
}
