% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rplanes_explorer}
\alias{rplanes_explorer}
\title{\code{rplanes} explorer app launcher}
\usage{
rplanes_explorer(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \link[shiny:runApp]{shiny::runApp}}
}
\value{
This function operates as a side-effect and starts the \code{rplanes} Shiny app.
}
\description{
The \code{rplanes} explorer app allows a user to interactively upload their own data (or view an internal example) to explore the plausibility analysis functionality.
}
\examples{
\dontrun{
# Launch the explorer app
rplanes_explorer(host = "0.0.0.0",
                 launch.browser = TRUE,
                 port = 80)
}
}
