% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_chunk}
\alias{to_chunk}
\title{Chunk a vector}
\usage{
to_chunk(x, size)
}
\arguments{
\item{x}{Vector to be split into chunks as large as teh "size" specified}

\item{size}{Width of the chunks for "x" vector}
}
\value{
A list with as many elements as the number of chunks created. Each element will include vector with a length equal to the "size" specified.
}
\description{
This unexported helper function creates a list with contents of a vector spit into chunks. The user can specify how large each chunk should be with the "size" argument.
}
