% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_location}
\alias{valid_location}
\title{Validate location}
\usage{
valid_location(location, input, seed)
}
\arguments{
\item{location}{Character vector with location code; the location must appear in input and seed}

\item{input}{Input signal data to be scored; object must be \link[=to_signal]{forecast}}

\item{seed}{Prepared \link[=plane_seed]{seed}}
}
\value{
The validation will return with a \code{stop()} if the location is not found in the seed or input signal. Otherwise the function will invisibly return \code{TRUE} indicating that the location is valid.
}
\description{
This unexported helper is used inside of the individual plausibility component functions (e.g., \code{plane_diff()}) to validate that the location specified appears in both the input signal and seed and that the location has as many values as other locations in the seed.
}
