% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epiweek_start}
\alias{epiweek_start}
\title{Resolve start of week from a given date}
\usage{
epiweek_start(date)
}
\arguments{
\item{date}{Date to be queried}
}
\value{
Date of the first day of the epiweek for the input date.
}
\description{
Resolve start of week from a given date
}
