% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_forecast}
\alias{is_forecast}
\title{Check if object is forecast}
\usage{
is_forecast(x)
}
\arguments{
\item{x}{Input object to be checked}
}
\value{
Logical as to whether or not the input object inherits the "signal" and "forecast" classes.
}
\description{
Check if object is forecast
}
