% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{month_start}
\alias{month_start}
\title{Resolve start of month from a given date}
\usage{
month_start(date)
}
\arguments{
\item{date}{Date to be queried}
}
\description{
Resolve start of month from a given date
}
