% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planes.R
\name{plane_repeat}
\alias{plane_repeat}
\title{Repeat component}
\usage{
plane_repeat(location, input, seed, tolerance = NULL, prepend = NULL)
}
\arguments{
\item{location}{Character vector with location code; the location must appear in input and seed}

\item{input}{Input signal data to be scored; object must be one of \link[=to_signal]{forecast}}

\item{seed}{Prepared \link[=plane_seed]{seed}}

\item{tolerance}{Integer value for the number of allowed repeats before flag is raised. Default is \code{NULL} and allowed repeats will be determined from seed.}

\item{prepend}{Integer value for the number of values from seed to add before the evaluated signal. Default is \code{NULL} and the number of values will be determined from seed.}
}
\value{
A \code{list} with the following values:
\itemize{
\item \strong{indicator}: Logical as to whether or not the value is repeated sequentially k number of times.
\item \strong{repeats}: A \code{tibble} with repeating values found. If there are no repeats (i.e., indicator is \code{FALSE}) then the \code{tibble} will have 0 rows.
}
}
\description{
This function evaluates whether consecutive values in observations or forecasts are repeated a k number of times. This function takes in a \link[=to_signal]{forecast} object that is either from an observed dataset or forecast dataset.
}
