% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{plane_seed}
\alias{plane_seed}
\title{driver to create seeds for every location in data}
\usage{
plane_seed(input, cut_date = NULL)
}
\arguments{
\item{input}{fixme}

\item{cut_date}{fixme}
}
\description{
driver to create seeds for every location in data
}
