% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane_shape_temp.R
\name{plane_shape}
\alias{plane_shape}
\title{Shape Component}
\usage{
plane_shape(location, input, seed)
}
\arguments{
\item{location}{Character vector with location code; the location must appear in input and seed}

\item{input}{Input signal data to be scored; object must be one of \link[=to_signal]{forecast} or \link[=to_signal]{observed}}

\item{seed}{Prepared \link[=plane_seed]{seed}}
}
\value{
A \code{list} with the following values:
\itemize{
\item \strong{indicator}: Logical as to whether or not the shape of the evaluated signal is novel
\item \strong{output}: An n x 5 tibble. The columns are:
\itemize{
\item \strong{date}: The dates corresponding to all observed and forecast data (formatted as date)
\item \strong{point}: The incidence of all observed and forecast data (e.g., hospitalization rates)
\item \strong{type}: Indicates whether the data row is observed or forecast data
\item \strong{indicator}: Logical as to whether or not the shape of the evaluated signal is novel
\item \strong{location}: A character vector with the location code
}
}
}
\description{
This function evaluates the shape of the trajectory of the forecast signal and compares that shape to existing shapes in observed data. If the shape is flagged as a novel shape, a flag is raised, and the signal is considered implausible. See details for further information.
}
\details{
This function uses a Dynamic Time Warping (DTW) algorithm to identify shapes within the observed seed data.

sliding windows
dtw on observed
dtw comparing observed distances to signal distances
if distance > threshold, than it is novel and it is flagged as implausible
}
\examples{
# We'll use the HHS Protect data that is internal to rplanes:

hosp <- read.csv(system.file("extdata/observed/hdgov_hosp_weekly.csv", package = "rplanes"))

tmp_hosp <-
 hosp \%>\%
 dplyr::select(date, location, flu.admits) \%>\%
 dplyr::mutate(date = as.Date(date))

 prepped_observed <- rplanes::to_signal(tmp_hosp, outcome = "flu.admits", type = "observed", resolution = "weeks")

 prepped_forecast <- rplanes::read_forecast(system.file("extdata/forecast/2022-10-31-SigSci-TSENS.csv", package = "rplanes")) \%>\%
   rplanes::to_signal(., outcome = "flu.admits", type = "forecast", horizon = 4)

prepped_seed <- rplanes::plane_seed(prepped_observed, cut_date = "2022-10-29")

# First, an example where the shape is novel and a flag is raised:

plane_shape(location = "13", input = prepped_forecast, seed = prepped_seed)

# Next, an example where a flag is not raised:

plane_shape(location = "06", input = prepped_forecast, seed = prepped_seed)

}
\references{
Toni Giorgino. Computing and Visualizing Dynamic Time Warping Alignments in R: The dtw Package. Journal of Statistical Software, 31(7), 1-24. doi:10.18637/jss.v031.i07

Tormene, P.; Giorgino, T.; Quaglini, S. & Stefanelli, M. Matching incomplete time series with dynamic time warping: an algorithm and an application to post-stroke rehabilitation. Artif Intell Med, 2009, 45, 11-34. doi:10.1016/j.artmed.2008.11.007
}
