% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planes.R
\name{plane_taper}
\alias{plane_taper}
\title{Taper component}
\usage{
plane_taper(location, input, seed)
}
\arguments{
\item{location}{Character vector with location code; the location must appear in input and seed}

\item{input}{Input signal data to be scored; object must be one of \link[=to_signal]{forecast}}

\item{seed}{Prepared \link[=plane_seed]{seed}}
}
\value{
A \code{list} with the following values:
\itemize{
\item \strong{indicator}: Logical as to whether or not the prediction interval width tapers with advancing horizons
\item \strong{widths}: Consecutive interval widths for
}
}
\description{
This function evaluates whether or not the evaluated signal interval tapers (i.e., decreases in width) as horizons progress. The interval used in this plausbility component is drawn from the upper and lower bounds of the forecasted prediction interval. As such, the only accepted signal format is \link[=to_signal]{forecast}, which will include upper and lower bounds.
}
