% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planes.R
\name{plane_taper}
\alias{plane_taper}
\title{Taper component}
\usage{
plane_taper(location, input, seed)
}
\arguments{
\item{location}{Character vector with location code; the location must appear in input and seed}

\item{input}{Input signal data to be scored; object must be one of \link[=to_signal]{forecast}}

\item{seed}{Prepared \link[=plane_seed]{seed}}
}
\value{
FIXME
}
\description{
FIXME
}
