% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_forecast}
\alias{read_forecast}
\title{Read in forecast file}
\usage{
read_forecast(file, pi_width = 95)
}
\arguments{
\item{file}{fixme}

\item{pi_width}{fixme}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item \strong{location}
\item \strong{date}
\item \strong{horizon}
\item \strong{lower}
\item \strong{point}
\item \strong{upper}
}
}
\description{
Read in forecast file
}
