% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplanes-package.R
\docType{package}
\name{rplanes-package}
\alias{rplanes}
\alias{rplanes-package}
\title{rplanes: Plausibility Analysis of Epidemiological Signals in 'R'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'rplanes' package provides functionality to prepare data and and analyze plausibility of both forecasted and reported epidemiological signals. The functions implement a set of plausbility algorithms that are agnostic to geographic and time resolutions and are calculated independently then presented as a combined score.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://signaturescience.github.io/rplanes/}
}

}
\author{
\strong{Maintainer}: VP Nagraj \email{pnagraj@signaturescience.com} (\href{https://orcid.org/0000-0003-0060-566X}{ORCID})

Authors:
\itemize{
  \item Desiree Williams (\href{https://orcid.org/0000-0002-0524-2055}{ORCID})
  \item Amy Benefield
}

}
\keyword{internal}
