% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{seed_engine}
\alias{seed_engine}
\title{helper function used in plane_seed()}
\usage{
seed_engine(.input, .location, .cut_date = NULL)
}
\arguments{
\item{.input}{fixme}

\item{.location}{fixme}

\item{.cut_date}{fixme}
}
\description{
helper function used in plane_seed()
}
