% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{to_signal}
\alias{to_signal}
\title{Title}
\usage{
to_signal(
  .input,
  .outcome,
  .type = "observed",
  .resolution = "weeks",
  .horizon = NULL
)
}
\arguments{
\item{.input}{needs location (FIPS); date (in YYYY-mm-dd format); outcome value}

\item{.outcome}{fixme}

\item{.type}{fixme}

\item{.resolution}{fixme}

\item{.horizon}{fixme}
}
\description{
Title
}
