% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_dates}
\alias{valid_dates}
\title{Check that date span is valid}
\usage{
valid_dates(seed_date, signal_date, resolution, warn_incomplete = FALSE)
}
\arguments{
\item{seed_date}{fixme}

\item{signal_date}{fixme}

\item{resolution}{fixme}

\item{warn_incomplete}{FIXME}
}
\value{
Invisible: a list of input values
}
\description{
Description FIXME
}
\details{
Details here FIXME
}
\examples{
seed_date <- as.Date("2023-03-08")
signal_date <- as.Date("2023-03-15")
valid_dates(seed_date = seed_date, signal_date = signal_date, resolution="weeks")
x <- try(valid_dates(seed_date = seed_date,
                          signal_date = signal_date,
                          resolution="days"), silent=TRUE)
x
x <- try(valid_dates(seed_date = seed_date,
                          signal_date = signal_date,
                          resolution="months"), silent=TRUE)
x
}
