% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_dates}
\alias{valid_dates}
\title{Check that date span is valid}
\usage{
valid_dates(seed_date, signal_date, resolution, warn_incomplete = FALSE)
}
\arguments{
\item{seed_date}{Last date available in seed object}

\item{signal_date}{First date available in signal object}

\item{resolution}{Character vector specifying the temporal resolution (e.g., "days", "weeks", "months")}

\item{warn_incomplete}{Logical as to whether or not the validation should warn for completeness of seed and signal; default is \code{FALSE}}
}
\value{
Invisible: a list of input values
}
\description{
Check that date span is valid
}
\examples{
seed_date <- as.Date("2023-03-08")
signal_date <- as.Date("2023-03-15")
valid_dates(seed_date = seed_date, signal_date = signal_date, resolution="weeks")
x <- try(valid_dates(seed_date = seed_date,
                          signal_date = signal_date,
                          resolution="days"), silent=TRUE)
x
x <- try(valid_dates(seed_date = seed_date,
                          signal_date = signal_date,
                          resolution="months"), silent=TRUE)
x
}
